+++ examples/deployment/server.rb
@@ -2,10 +2,10 @@
 # frozen_string_literal: true
 
 require "async/http"
-require "async/service/managed/service"
-require "async/service/managed/environment"
+require "async/service/managed_service"
+require "async/service/managed_environment"
 
-class WebService < Async::Service::Managed::Service
+class WebService < Async::Service::ManagedService
 	def start
 		super
 		@endpoint = @evaluator.endpoint
@@ -31,7 +31,7 @@ class WebService < Async::Service::Managed::Service
 end
 
 module WebEnvironment
-	include Async::Service::Managed::Environment
+	include Async::Service::ManagedEnvironment
 	
 	def endpoint
 		Async::HTTP::Endpoint.parse("http://0.0.0.0:3000")