+++ test/async/service/managed_environment.rb
@@ -4,13 +4,13 @@
 # Copyright, 2025, by Samuel Williams.
 
 require "async/service/configuration"
-require "async/service/managed/environment"
+require "async/service/managed_environment"
 
-describe Async::Service::Managed::Environment do
+describe Async::Service::ManagedEnvironment do
 	let(:configuration) do
 		Async::Service::Configuration.build do
 			service "test-managed" do
-				include Async::Service::Managed::Environment
+				include Async::Service::ManagedEnvironment
 				
 				count 3
 				health_check_timeout 10
@@ -31,7 +31,7 @@ describe Async::Service::Managed::Environment do
 	it "provides default count as nil" do
 		configuration = Async::Service::Configuration.build do
 			service "test-default" do
-				include Async::Service::Managed::Environment
+				include Async::Service::ManagedEnvironment
 			end
 		end
 		
@@ -42,7 +42,7 @@ describe Async::Service::Managed::Environment do
 	it "provides default health_check_timeout as 30" do
 		configuration = Async::Service::Configuration.build do
 			service "test-default" do
-				include Async::Service::Managed::Environment
+				include Async::Service::ManagedEnvironment
 			end
 		end
 		
@@ -57,7 +57,7 @@ describe Async::Service::Managed::Environment do
 	it "compacts nil values from container_options" do
 		configuration = Async::Service::Configuration.build do
 			service "test-nil" do
-				include Async::Service::Managed::Environment
+				include Async::Service::ManagedEnvironment
 				
 				count nil
 				health_check_timeout nil