+++ releases.md
@@ -1,5 +1,11 @@
 # Releases
 
+## Unreleased
+
+  - Renamed `Async::Service::Generic` -> `Async::Service::GenericService`, added compatibilty alias.
+  - Renamed `Async::Service::Managed::Service` -> `Async::Service::ManagedService`.
+  - Renamed `Async::Service::Managed::Environment` -> `Async::Service::ManagedEnvironment`.
+
 ## v0.15.1
 
   - `Managed::Service` should run within `Async do ... end`.